+++ lib/array2d.dart
@@ -1,3 +1,2 @@
 export 'package:array2d/src/array2d.dart';
 export 'package:array2d/src/point_data.dart';
-export 'package:array2d/src/column_viewer.dart';